Fix implementation to respect 'output' argument
Earlier, the script didn't consider the 'output' argument at
all. Instead it assumed the output path by suffixing '.extract' and
'.recompressed' for extracting and compressing respectively. This commit
fixing it by considering the output path as specified in the argument
which makes it more explicit. Also added a validation to check that
input and output paths are not the same.