Improve converter selection using fnmatch
This commit makes the implementation for choosing converter more
flexible by using fnmatch. The CONVERTERS map is replaced with list of
pairs (tuples) where every pair consists of a valid fnmatch pattern and
a converter instance.

When selecting the correct converter for a path, the path is compared
against all the patterns (in order) using fnmatch and the first matching
pattern is selected. If no patterns match, NoopConverter is used in
order to prevent any loss of information.